+++ packages/apollo-server-core/tsconfig.json
@@ -12,6 +12,7 @@
     { "path": "../apollo-engine-reporting" },
     { "path": "../apollo-server-caching" },
     { "path": "../apollo-server-errors" },
+    { "path": "../apollo-server-plugin-base" },
     { "path": "../apollo-tracing" },
     { "path": "../graphql-extensions" }
   ]
