+++ packages/apollo-server-core/tsconfig.requestPipelineAPI.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.base",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./dist"
+  },
+  "files": ["src/requestPipelineAPI.ts"],
+  "references": []
+}
