+++ packages/apollo-server-plugin-base/src/index.ts
@@ -1,35 +1,10 @@
-import { GraphQLSchema } from 'graphql';
-import { KeyValueCache } from 'apollo-server-caching';
+import { GraphQLRequest } from 'apollo-server-core/src/requestPipelineAPI';
 
-export abstract class ApolloServerPluginBase {
-  [key: string]: any;
-  constructor(public options?: any) {}
-  serverWillStart?(parent: any): void;
+export abstract class ApolloServerPlugin {
+  serverWillStart?(): void;
+  requestDidStart?(): GraphQLRequestListener;
 }
 
-interface ApolloServerRequestListenerInterface {
-  start?(): void;
+export interface GraphQLRequestListener {
+  prepareRequest?(request: GraphQLRequest): void;
 }
-
-export abstract class ApolloServerRequestListenerBase
-  implements ApolloServerRequestListenerInterface {
-  constructor() {}
-  start?(): void;
-}
-
-interface PluginEventBase {
-  name: string;
-  args?: any;
-}
-
-export interface PluginEventServerWillStart extends PluginEventBase {
-  args: {
-    schema: GraphQLSchema;
-    persistedQueries?: {
-      cache: KeyValueCache;
-    };
-    engine?: any;
-  };
-}
-
-export type PluginEvent = PluginEventServerWillStart;
